+++ modules/Student_Billing/Menu.php
@@ -5,7 +5,7 @@
  * @uses $menu global var
  *
  * @see  Menu.php in root folder
- * 
+ *
  * @package RosarioSIS
  * @subpackage modules
  */
@@ -20,13 +20,11 @@ $menu['Student_Billing']['admin'] = array(
 	1 => _( 'Reports' ),
 	'Student_Billing/StudentBalances.php' => _( 'Student Balances' ),
 	'Student_Billing/DailyTransactions.php' => _( 'Daily Transactions' ),
-	'Student_Billing/DailyTotals.php' => _( 'Daily Totals' ),
 	'Student_Billing/Statements.php' => _( 'Print Statements' ),
 );
 
 // FJ fix error Warning: Invalid argument supplied for foreach().
-$menu['Student_Billing']['teacher'] = array(
-);
+$menu['Student_Billing']['teacher'] = array();
 
 $menu['Student_Billing']['parent'] = array(
 	'title' => _( 'Student Billing' ),
@@ -39,5 +37,4 @@ $menu['Student_Billing']['parent'] = array(
 	'Student_Billing/Statements.php&_ROSARIO_PDF' => _( 'Print Statements' ),
 );
 
-$exceptions['Student_Billing'] = array(
-);
+$exceptions['Student_Billing'] = array();